+++ packages/protocol/src/index.ts
@@ -1,0 +1,35 @@
+export {
+  PROTOCOL_VERSION,
+  EVENT_TYPES,
+  type TeamMeta,
+  type RunSummary,
+  type RunEvent,
+  type RunStartEvent,
+  type SceneStartEvent,
+  type ActionStartEvent,
+  type ActionEndEvent,
+  type AssertionEvent,
+  type WarningEvent,
+  type SceneEndEvent,
+  type RunProgressEvent,
+  type RunEndEvent,
+} from './events.js'
+
+export {
+  COMMAND_TYPES,
+  type Command,
+  type RunReplayCommand,
+  type RunStopCommand,
+  type RunPauseCommand,
+  type RunResumeCommand,
+} from './commands.js'
+
+export {
+  isEventShaped,
+  isRunEvent,
+  isCommand,
+  encodeEvent,
+  decodeEvent,
+  encodeCommand,
+  decodeCommand,
+} from './codec.js'